portfolio

+++ database/migrations/2023_03_22_212642_create_data_sets_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->string('symbol');
             $table->date('date');
-            $table->decimal('price');
+            $table->float('price', 10, 5);
             $table->timestamps();
         });
     }
